feat(api): Add user info to order placement
User information is required to create an order, otherwise the order is
not a legitimate one.

In addition the return object from the orders service was reduced
because the entire order does not need to be returned, as an order Id
can be used to get the entire order. Separation of concerns via
different REST resources is more important than unneeded data.